Move XGM mock devices to their own module in .mockdata
